Output job name only after it has been formatted with new member id (#14)
This fixes an (aesthetic) issue with the output of the _run-ensemble_
command. The code outputs the sbatch command it will use for running a
specific ensemble, however the name of the runscript is wrong. In more
detail, the name of the runscript corresponds to the one of previous
member, i.e. the numbers are shifted by 1. Note that this is only an
output problem.

This change simply moves the output of the "sbatch command" string to after
the `job` variable is formatted with the correct member id.

Commit by: Mikael Stellio