Update the tests to cover the new edge cases.
The test data is changed to a much better controlled data set instead
of random values.

Tolerances are now smaller than 1 which is much more realistic and
allows for testing small numbers.

New tests are added to make sure that the test only validates if both
dataframes are close to zero.