Consolidate probtest (#20)
Same as old merge request:
https://github.com/MeteoSwiss/probtest/pull/12 but without using fork.

As part of the effort to [Make probtest tolerance test more
robust](https://gitlab.dkrz.de/icon/icon/-/issues/9476) and the [C2SM
task to consolidate
probtest](https://github.com/C2SM/Tasks-Support/issues/2), the following
features have been adapted or added. The task will be continued after
this merge but the current changes are needed to fix the tolerance test
for the [GPU port of
lseaice](https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/824) as
more ensemble members are needed for the test to pass. This merge
request should follow the merge of [Consolidate probtest (C2SM
task)](https://gitlab.dkrz.de/icon/icon-nwp/-/merge_requests/980) on
icon-nwp, which depends on those changes.
- Probtest now fails when there is no overlap between the variables in
the reference and the test file. It gives a warning when are there
variables which are not in both the reference and the test case. It also
specifies which variables those are.
- Probtest now checks if the number of time steps is equal in the
reference and test case. If they are not equal, probtest gives a warning
and checks only the overlapping time steps
- The `check_plot.py` function has been changed so that the fields with
the largest error relative to tolerance instead of the largest
difference are plotted first (failing fields may not be plotted
otherwise)
- Instead of `--member_ids` the input options `--member-num` and
`--member-type` were implemented.
- `--member-num` sets the number of ensemble members. If a single number
x is given, the member numbers are 1, 2,...,x, another input option to
set the member numbers by separating numbers with comma, e.g. 2,4,7,10.
(default is 10). Each number is connected to a seed, which is set within
the `seeds.txt` file. In includes 100 pre-defined seeds.
- `--member-type` is only important for BB, where it is set to `mixed`
or `double`. BB would otherwise create scripts with the same name for
double and mixed precision and would thus overwrite some. The default is
`""`.
- The seeds are now set with help of the `seeds.txt` file instead of
being created from the hash of the member id. Specific seeds can be
selected by setting the member_num acordingly (see above).
- Added new functions to unit test
- Updated README: improved description for developing in probtest and
fixed links.
- Fixed unittests: added file_ID to the created dataframes for testing
to use same format as is being used in probtest.

---------

Co-authored-by: Marek Jacob <MeraX@users.noreply.github.com>