Remove Account from sbatch submit_command (#29)
Setting the account for the `sbatch` in the `submit_command` explicitly
to `g110` is not very helpful when running probtest as a user that is
not part of g110 or running it on another system. I think this parameter
should not be required on any system and we can simply remove it.

The `submit_command` is only used in `run_ensemble`. Buildbot, however,
uses `run_ensemble` only in the `dry` mode, i.e. no jobs are submitted.
Thus `submit_command` is not used by Buildbot at all.