Overhaul ci (#31)
Replace `e2e` tests with a new set of tests that is independent of ICON
data. These tests basically tests the command-line interfaces of the
click-subcommand. So far `cdo-table`, `tolerance`, `perturb`,
`performance` and `stats` are covered.

### Pytest fixtures
In order to handle data requirements of the tests a set of fixtures is
used.
This allows to organize the data requirements of the tests in a central
place.

### Execute all tests using Pytest
In order to start transition to pytests, collect and execute all tests
using pytest

### Reference files in `tests/data`
The tests compare its results to precomputed values located in the repo.
This ensures consistency over time. In case new reference data is
required, pytests automatically stores
the new references for each test-session in a folder in `/tmp`. The
location is printed to terminal
when running pytest.

### cdo-table overrides function at runtime
See https://github.com/MeteoSwiss/probtest/issues/32,
a quick and dirty fix has been introduce in the PR.

---------

Co-authored-by: Jonas Jucker <jonas.jucker@c2sm.ethz.ch>