Revise member selection (#54)
- clarify naming convenction, for `member_num` use either `member_id`,
`member_ids`, or `member_count`
- revise the member selection algorithm, instead of checking random it
uses this deterministic approach

At each step, every ensemble member is evaluated to determine how many
of the remaining members satisfy the given tolerances when compared to
it. The ensemble member that results in the highest number of other
members passing the tolerances is selected. This process is repeated
until either all members are successfully validated using the current
tolerances, or a maximum number of attempts is reached. In the latter
case, the tolerance factor is increased, and the process continues.

---------

Co-authored-by: Annika Lauber <annika.lauber@c2sm.ethz.ch>